Factor out common part from protractor confs
 - add firefox protractor conf
