+++ etc/protractorConfSauce.js
@@ -1,16 +1,9 @@
-var exec = require("sync-exec");
-var fs = require("fs");
-var ini = require("ini");
 var pr = process.env.TRAVIS_PULL_REQUEST;
 var name = ((pr === "false") ? "" : "#" + pr + " ") + process.env.TRAVIS_COMMIT;
 
-exports.config = {
-    suites: {
-        // FIXME: mercator tests fail on travis
-        //current: "../src/current/current/tests/acceptance/*Spec.js",
-        core: "../src/adhocracy_frontend/adhocracy_frontend/tests/acceptance/*Spec.js"
-    },
-    baseUrl: "http://localhost:9090",
+var common = require("./protractorCommon.js");
+
+var local = {
     sauceUser: "liqd",
     sauceKey: "77600374-1617-4d7b-b1b6-9fd82ddfe89c",
 
@@ -20,26 +13,6 @@ exports.config = {
         "build": process.env.TRAVIS_BUILD_NUMBER,
         "name": name
     },
-    beforeLaunch: function() {
-        exec("bin/supervisord");
-        exec("bin/supervisorctl restart adhocracy_test:test_zeo test_backend_with_ws adhocracy_test:test_autobahn adhocracy_test:test_frontend");
-        exec("src/current/current/tests/acceptance/setup_test.sh");
-    },
-    afterLaunch: function() {
-        exec("bin/supervisorctl stop adhocracy_test:test_zeo test_backend_with_ws adhocracy_test:test_autobahn adhocracy_test:test_frontend");
-        exec("rm -rf var/test_zeodata/Data.fs* var/test_zeodata/blobs");
-    },
-    onPrepare: function() {
-        var getMailQueuePath = function() {
-            var testConf = ini.parse(fs.readFileSync("etc/test_with_ws.ini", "utf-8"));
-            return testConf["app:main"]["mail.queue_path"]
-                   .replace("%(here)s", process.cwd() + "/etc");
-        };
-
-        browser.params.mail = {
-            queue_path: getMailQueuePath()
-        }
-    },
     allScriptsTimeout: 21000,
     getPageTimeout: 20000,
     jasmineNodeOpts: {
@@ -48,4 +21,6 @@ exports.config = {
         isVerbose: true,
         includeStackTrace: true
     }
-}
+};
+
+exports.config = Object.assign({}, common.config, local);
